Update patch set 1
Patch Set 1:

As for the branch naming convention, it is already used elsewhere so it is not gonna change. But when the content is reviewed and tested it will go back to master.

I could have postponed this patch for the webserver, but as we are to install the mailing-lists very very soon, then this becomes compulsory.

As for the dependency, i made a mistake. the 'main' entrypoint is ok with this usage, so I can leave it. Will send a fix soon.

Patch-set: 1
Label: Code-Review=0