Update patch set 1
Patch Set 1:

All hosts with roles from OSPO/OSCI depends on EPEL and it's installed in the playbook in the main repo.

The doc was updated to tell this requirement in a previous PR but this is not this role's function to deal with this.

This role won't work with this change and previously koji was (unfortunately) installed manually so this would not have worked on EL7 either.

Patch-set: 1
Reviewer: Gerrit User 422 <422@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0